requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
narwhals==1.29.1
numpy==1.23.5
packaging==24.2
pandas==1.5.3
pandas_ta==0.3.14b0
patsy==1.0.1
pillow==11.1.0
plotly==6.0.0
plotting==0.0.7
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
statsmodels==0.14.4
tzdata==2025.1
urllib3==2.3.0
utils==1.0.2
